now: dont cast directly from i64 to u128
seems it gets sign extended :(

+++ src/now.rs
@@ -108,7 +108,7 @@ pub fn run(args: &Args) {
         }
         Action::Convert { timestamp } => {
             let unix = timestamp.timestamp();
-            let ob = fe(unix as u128, args.seed);
+            let ob = fe(u128::from(unix as u64), args.seed);
             println!("{}", b32(ob));
         }
         Action::Decode { blob } => {
